+++ gogod/telegram.py
@@ -1,0 +1,116 @@
+#!/usr/bin/python
+
+import os, sys
+import time
+import consolelog
+import telepot
+import config
+import threading
+import requests
+
+_rate_limit_telegram    = 0.25 #seconds
+APPLICATION_PATH        = os.path.abspath(os.path.dirname(sys.argv[0]))
+LOG_TITLE               = "Telegram"
+
+
+class TelegramBot():
+    def __init__(self, gogod_config=None, sendKeyValue=None):
+        self.last_handle_time   = 0
+        self.conf               = config.Config() if gogod_config is None else gogod_config
+        self.sendKeyValue       = sendKeyValue
+        self.token            = self.conf.get_telegram_bot_token()
+        self.latest_sender_id   = self.get_sender()
+        self.bot                = None
+        self.bot_worker         = None
+        consolelog.log(LOG_TITLE, "Created Class")
+
+        self.connect()
+
+    def getAPIKey(self):
+        self.token = self.conf.get_telegram_bot_token()
+
+    def connect(self):
+
+        self.getAPIKey()
+        if not self.token:
+            return
+
+        if self.bot is not None:
+            self.bot        = None
+
+        self.bot = telepot.Bot(self.token)
+        try:
+            self.bot.getMe()
+            consolelog.log(LOG_TITLE, "Connected")
+        except:
+            self.bot = None
+            consolelog.log(LOG_TITLE, "Invalid token")
+            return
+
+        self.bot.message_loop({'chat': self.on_telegram_message})
+
+    def handle_gogo_message(self, message):
+        if (time.time() - self.last_handle_time) < _rate_limit_telegram:
+            return False
+        self.last_handle_time = time.time()
+
+        messages = message.split(',')
+        if len(messages) == 1:
+            self.send_telegram_message(message)
+        elif len(messages) > 1:
+            if messages[0] == 'image' or messages[0] == 'picture' or messages[0] == 'photo':
+                thread_send_photo = threading.Thread(target=self.send_telegram_photo, args=(messages[1],))
+                thread_send_photo.start()
+            elif messages[0] == 'message':
+                self.send_telegram_message(messages[1])
+            else:
+                self.send_telegram_message(message)
+
+    def on_telegram_message(self, msg):
+        content_type, chat_type, chat_id = telepot.glance(msg)
+        message = msg['text']
+    
+        if self.latest_sender_id != chat_id:
+            self.latest_sender_id = chat_id
+            self.save_sender(chat_id)
+
+        if self.bot is None:
+            return
+
+        self.bot.sendMessage(chat_id, message)
+        consolelog.log(LOG_TITLE, "message : %s" % message)
+        if self.sendKeyValue is not None:
+            self.sendKeyValue("telegram", message)
+
+    def send_telegram_message(self, message):
+        if self.latest_sender_id is not None and self.bot is not None:
+            self.bot.sendMessage(self.latest_sender_id, text=message)
+
+    def send_telegram_photo(self, filename):
+        if self.latest_sender_id is None or self.bot is None:
+            return
+
+        filepath = self.conf.auto_filepath_image(filename)
+        if not os.path.exists(filepath):
+            return
+        consolelog.log(LOG_TITLE, 'sending image')
+        url     = 'https://api.telegram.org/bot%s/sendPhoto' % self.token
+        files   = {'photo': open(filepath, 'rb')}
+        data    = {'chat_id': self.latest_sender_id,'caption':filename}
+        try:
+            r       = requests.post(url, data=data, files=files)
+            consolelog.log(LOG_TITLE, 'sent result %s' % r.json()['ok'])
+        except:
+            consolelog.log(LOG_TITLE, 'sent image fail')
+
+
+    def on_gogo_message(self, title, message):
+        if title == "telegram":
+            print "%s = %s" %(title, message)
+            self.send_telegram_message( "%s" % (message) )
+
+    def save_sender(self, sender_id):
+        self.conf.save(self.conf.telegram_bot_sender, sender_id)
+
+    def get_sender(self):
+        return self.conf.get(self.conf.telegram_bot_sender)